+++ src/xrunDebug.ts
@@ -158,7 +158,7 @@ export class XrunDebugSession extends LoggingDebugSession {
 		// make VS Code support data breakpoints
 		response.body.supportsDataBreakpoints = true;
 
-		response.body.supportsConditionalBreakpoints = false;
+		response.body.supportsConditionalBreakpoints = true;
 
 		response.body.supportsHitConditionalBreakpoints = true;
 
@@ -252,7 +252,7 @@ export class XrunDebugSession extends LoggingDebugSession {
 
 		// set and verify breakpoint locations
 		const actualBreakpoints0 = (args.breakpoints || []).map(client_bp => {
-			const runtime_bp = this._runtime.setBreakPoint(path, client_bp.line, client_bp.hitCondition);
+			const runtime_bp = this._runtime.setBreakPoint(path, client_bp.line, client_bp.hitCondition, client_bp.condition);
 			const bp: DebugProtocol.Breakpoint = new Breakpoint(runtime_bp.verified, runtime_bp.line);
 			bp.id = runtime_bp.id;
 			return bp;